additionalAC param for inputbar + move lib
