+++ front/components/assistant/conversation/input_bar/InputBar.tsx
@@ -45,6 +45,7 @@ export function AssistantInputBar({
   onSubmit,
   conversationId,
   stickyMentions,
+  additionalAgentConfigurations,
 }: {
   owner: WorkspaceType;
   onSubmit: (
@@ -54,6 +55,7 @@ export function AssistantInputBar({
   ) => void;
   conversationId: string | null;
   stickyMentions?: AgentMention[];
+  additionalAgentConfigurations?: AgentConfigurationType[];
 }) {
   const [contentFragmentBody, setContentFragmentBody] = useState<
     string | undefined
@@ -65,6 +67,8 @@ export function AssistantInputBar({
     workspaceId: owner.sId,
     agentsGetView: conversationId ? { conversationId } : "list",
   });
+  agentConfigurations.push(...(additionalAgentConfigurations ?? []));
+
   const sendNotification = useContext(SendNotificationsContext);
 
   const [isAnimating, setIsAnimating] = useState<boolean>(false);
@@ -265,6 +269,7 @@ export function FixedAssistantInputBar({
   onSubmit,
   stickyMentions,
   conversationId,
+  additionalAgentConfigurations,
 }: {
   owner: WorkspaceType;
   onSubmit: (
@@ -274,6 +279,7 @@ export function FixedAssistantInputBar({
   ) => void;
   stickyMentions?: AgentMention[];
   conversationId: string | null;
+  additionalAgentConfigurations?: AgentConfigurationType[];
 }) {
   return (
     <div className="4xl:px-0 fixed bottom-0 left-0 right-0 z-20 flex-initial lg:left-80">
@@ -283,6 +289,7 @@ export function FixedAssistantInputBar({
           onSubmit={onSubmit}
           conversationId={conversationId}
           stickyMentions={stickyMentions}
+          additionalAgentConfigurations={additionalAgentConfigurations}
         />
       </div>
     </div>
