+++ front/components/assistant/conversation/lib.ts
